requirements.txt
antiorm==1.2.1
click==8.0.1
colorama==0.4.4
db==0.1.1
db-sqlite3==0.0.1
Flask==2.0.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
PyJWT==1.4.2
pytz==2021.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
